Move auto-configuration and properties to starters
