+++ sso-kit-starter-hilla/src/main/java/dev/hilla/sso/starter/SingleSignOnContext.java
@@ -28,8 +28,6 @@ import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import com.vaadin.sso.core.SingleSignOnProperties;
-
 import dev.hilla.Nonnull;
 import jakarta.servlet.http.HttpServletRequest;
 import reactor.core.publisher.Flux;
