+++ sso-kit-starter/src/main/java/com/vaadin/sso/starter/SingleSignOnDefaultBeans.java
@@ -7,7 +7,7 @@
  * See <https://vaadin.com/commercial-license-and-service-terms> for the full
  * license.
  */
-package com.vaadin.sso.core;
+package com.vaadin.sso.starter;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
