+++ sso-kit-starter/src/test/java/com/vaadin/sso/starter/SingleSignOnConfigurationTest.java
@@ -20,8 +20,6 @@ import org.springframework.security.oauth2.client.registration.ClientRegistratio
 
 import com.vaadin.flow.spring.SpringBootAutoConfiguration;
 import com.vaadin.flow.spring.SpringSecurityAutoConfiguration;
-import com.vaadin.sso.core.SingleSignOnDefaultBeans;
-import com.vaadin.sso.core.SingleSignOnProperties;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
